# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Fleet Operations API
  version: 1.0.0
  description: 车队经营管理系统统一 API 契约。详细业务规则以项目需求基线文档为准。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: 客户分页列表
      x-permission: route:list
      responses: {'200': {description: OK}}
    post:
      tags: [Customers]
      summary: 新增客户
      x-permission: route:create
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CustomerRequest'}}}}
      responses: {'200': {description: OK}, '409': {$ref: '#/components/responses/BusinessConflict'}}
  /customers/{id}:
    put:
      tags: [Customers]
      summary: 编辑客户
      x-permission: route:edit
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CustomerRequest'}}}}
      responses: {'200': {description: OK}}
  /customers/{id}/status:
    put:
      tags: [Customers]
      summary: 启用/停用客户
      x-permission: route:edit
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/StatusRequest'}}}}
      responses: {'200': {description: OK}}
  /products:
    get:
      tags: [Products]
      summary: 产品分页列表
      x-permission: route:list
      responses: {'200': {description: OK}}
    post:
      tags: [Products]
      summary: 新增产品
      x-permission: route:create
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProductRequest'}}}}
      responses: {'200': {description: OK}, '409': {$ref: '#/components/responses/BusinessConflict'}}
  /products/{id}:
    put:
      tags: [Products]
      summary: 编辑产品
      x-permission: route:edit
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProductRequest'}}}}
      responses: {'200': {description: OK}}
  /products/{id}/status:
    put:
      tags: [Products]
      summary: 启用/停用产品
      x-permission: route:edit
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/StatusRequest'}}}}
      responses: {'200': {description: OK}}
...
  /drivers/{id}:
    get:
      tags: [Drivers]
      summary: 司机详情
      x-permission: driver:list
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
    put:
      tags: [Drivers]
      summary: 编辑司机
      x-permission: driver:edit
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
  /drivers/{id}/bind-vehicle:
    post:
      tags: [Drivers]
      summary: 绑定车辆
      x-permission: driver:bindVehicle
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}, '409': {$ref: '#/components/responses/BusinessConflict'}}
  /drivers/{id}/unbind-vehicle:
    post:
      tags: [Drivers]
      summary: 解绑车辆
      x-permission: driver:unbindVehicle
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}, '409': {$ref: '#/components/responses/BusinessConflict'}}
...
  /vehicles/{id}:
    get:
      tags: [Vehicles]
      summary: 车辆详情
      x-permission: vehicle:view
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
    put:
      tags: [Vehicles]
      summary: 编辑车辆
      x-permission: vehicle:edit
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
    delete:
      tags: [Vehicles]
      summary: 删除车辆
      x-permission: vehicle:delete
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
  /vehicles/{id}/binding-history:
    get:
      tags: [Vehicles]
      summary: 车辆绑定历史
      x-permission: vehicle:bindingHistory
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
...
  /trailers/{id}:
    put:
      tags: [Trailers]
      summary: 编辑车挂
      x-permission: trailer:edit
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
  /trailers/{id}/bind-vehicle:
    post:
      tags: [Trailers]
      summary: 绑定车辆
      x-permission: trailer:bindVehicle
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}, '409': {$ref: '#/components/responses/BusinessConflict'}}
  /trailers/{id}/unbind-vehicle:
    post:
      tags: [Trailers]
      summary: 解绑车辆
      x-permission: trailer:unbindVehicle
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}, '409': {$ref: '#/components/responses/BusinessConflict'}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: integer, format: int64}}
...
  responses:
    BusinessConflict:
      description: 业务规则冲突
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiResponse'}
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code: {type: string, examples: ['0']}
        message: {type: string}
        data: {}
        requestId: {type: string}
...
    StatusRequest:
      type: object
      required: [status]
      properties:
        status: {type: integer, enum: [0, 1]}
    CustomerRequest:
      type: object
      required: [customerName, status]
      properties:
        customerCode: {type: [string, 'null']}
        customerName: {type: string}
        status: {type: integer, enum: [0, 1]}
        remark: {type: [string, 'null']}
...
    ProductRequest:
      type: object
      required: [productName, status]
      properties:
        productCode: {type: [string, 'null']}
        productName: {type: string}
        status: {type: integer, enum: [0, 1]}
        remark: {type: [string, 'null']}
...